Masonry: move files to /Masonry directory
