+++ packages/gestalt/src/Masonry/MeasurementStore.js
@@ -1,6 +1,7 @@
 // @flow strict
 import { type Cache } from './Cache.js';
 
+// replace V with "number"
 export default class MeasurementStore<T: { ... } | $ReadOnlyArray<mixed>, V>
   implements Cache<T, V>
 {
