+++ packages/gestalt/src/Masonry/mindex.js
@@ -1,0 +1,11 @@
+// @flow strict
+
+export default function mindex(arr: $ReadOnlyArray<number>): number {
+  let idx = 0;
+  for (let i = 0; i < arr.length; i += 1) {
+    if (arr[i] < arr[idx]) {
+      idx = i;
+    }
+  }
+  return idx;
+}
