+++ packages/gestalt/src/Masonry/types.js
@@ -1,0 +1,8 @@
+// @flow strict
+
+export type Position = {|
+  top: number,
+  left: number,
+  width: number,
+  height: number,
+|};
