+++ packages/gestalt/src/ScrollFetch.js
@@ -1,8 +1,8 @@
 // @flow strict
 import { type Node, PureComponent } from 'react';
 import FetchItems from './FetchItems.js';
-import ScrollContainer from './ScrollContainer.js';
-import { getElementHeight, getScrollHeight, getScrollPos } from './scrollUtils.js';
+import ScrollContainer from './Masonry/ScrollContainer.js';
+import { getElementHeight, getScrollHeight, getScrollPos } from './Masonry/scrollUtils.js';
 import throttle, { type ThrottleReturn } from './throttle.js';
 
 type Props = {|
